+++ app/src/main/java/com/praca/dyplomowa/android/views/ProfileJobListFragmentView.kt
@@ -15,6 +15,7 @@ import com.praca.dyplomowa.android.databinding.FragmentProfileJobListViewBinding
 import com.praca.dyplomowa.android.utils.*
 import com.praca.dyplomowa.android.viewmodels.ProfileJobListViewModel
 import com.praca.dyplomowa.android.views.adapters.JobAdapter
+import com.praca.dyplomowa.android.views.adapters.JobHoursAdapter
 
 class ProfileJobListFragmentView : Fragment() {
 
@@ -22,6 +23,7 @@ class ProfileJobListFragmentView : Fragment() {
     private val binding get() = _binding!!
     private lateinit var viewModelProfileJobList: ProfileJobListViewModel
     private lateinit var jobAdapter: JobAdapter
+    private lateinit var jobHoursAdapter: JobHoursAdapter
     private lateinit var dateRange: DateRange
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -40,17 +42,15 @@ class ProfileJobListFragmentView : Fragment() {
         setObserverForError()
         setObserverForGetCompletedJobsAppliedToUser()
         setObserverForGetTodoJobsAppliedToUser()
-        setObserverForGetJobsForSpecifiedMonthAndUserAndCheckCompleted()
+        setObserverForGetJobsForSpecifiedMonthAndUser()
         setObserverForDeleteJob()
 
         if(arguments?.getString("dateRange") != null){
             dateRange = Gson().fromJson(arguments?.getString("dateRange")!!, DateRange::class.java)
         }
-
         binding.textViewTitleJobsListFragmentProfileView.setText(arguments?.getString("title"))
         binding.recyclerViewJobsListFragmentProfileView.layoutManager = LinearLayoutManager(requireContext())
-        jobAdapter = JobAdapter(recyclerViewJobsUtilsInterface)
-        binding.recyclerViewJobsListFragmentProfileView.adapter = jobAdapter
+
         getJobsAndUpdateRecyclerData()
 
         requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, object : OnBackPressedCallback(true) {
@@ -80,9 +80,9 @@ class ProfileJobListFragmentView : Fragment() {
         }
     }
 
-    private fun setObserverForGetJobsForSpecifiedMonthAndUserAndCheckCompleted(){
-        viewModelProfileJobList.jobResult.observe(viewLifecycleOwner){
-            jobAdapter.setupData(it.collection.toList())
+    private fun setObserverForGetJobsForSpecifiedMonthAndUser(){
+        viewModelProfileJobList.jobHoursResult.observe(viewLifecycleOwner){
+            jobHoursAdapter.setupData(it.collection.toList())
         }
     }
 
@@ -103,17 +103,30 @@ class ProfileJobListFragmentView : Fragment() {
 
     fun getJobsAndUpdateRecyclerData(){
         if(arguments?.getString("title") == resources.getString(R.string.textview_list_jobs_completed)){
+            setupJobAdapter()
             viewModelProfileJobList.getCompletedJobsAppliedToUser(SessionManager.getCurrentUserUsername(requireContext())!!)
         }else if(arguments?.getString("title") == resources.getString(R.string.textview_list_jobs_todo)){
+            setupJobAdapter()
             viewModelProfileJobList.getTodoJobsAppliedToUser(SessionManager.getCurrentUserUsername(requireContext())!!)
         }else{
-            viewModelProfileJobList.getJobsForSpecifiedMonthAndUserAndCheckCompleted(
+            setupJobHoursAdapter()
+            viewModelProfileJobList.getJobsForSpecifiedMonthAndUser(
                 startLong = dateRange.startLong,
                 endLong = dateRange.endLong,
                 username = arguments?.getString("username")!!
             )
         }
 
+    }
+
+    private fun setupJobAdapter(){
+        jobAdapter = JobAdapter(recyclerViewJobsUtilsInterface)
+        binding.recyclerViewJobsListFragmentProfileView.adapter = jobAdapter
+    }
+
+    private fun setupJobHoursAdapter(){
+        jobHoursAdapter = JobHoursAdapter(recyclerViewJobsUtilsInterface)
+        binding.recyclerViewJobsListFragmentProfileView.adapter = jobHoursAdapter
     }
 
     private val recyclerViewJobsUtilsInterface: RecyclerViewJobsUtilsInterface = object :
@@ -129,29 +142,31 @@ class ProfileJobListFragmentView : Fragment() {
         }
 
         override fun onLongClick(string: String) {
-            val materialDialog = MaterialAlertDialogBuilder(requireContext())
-                .setTitle(R.string.dialog_joblist_text_title)
-                .setMessage(R.string.dialog_joblist_text_message)
-                .setPositiveButton(R.string.dialog_joblist_positive_title) { dialog, which ->
-                    FragmentNavigationUtils.addFragmentOpenWithOneStringBundleValueAndSourceFragment(
-                        fragmentManager = parentFragmentManager,
-                        fragment = JobApplyToFragmentView(),
-                        argumentKey = "jobId",
-                        argumentValue = string,
-                        argumentSourceFragmentName = "ProfileJobListFragmentView"
-                    )
-                }
-                .setNeutralButton(R.string.dialog_joblist_neutral_title) { dialog, which ->
-                    dialog.dismiss()
-                }
-                .setNegativeButton(R.string.dialog_joblist_negative_title) { dialog, which ->
-                    viewModelProfileJobList.deleteJob(string)
-                }
+            if(SessionManager.getIsAdmin(requireContext())){
+                val materialDialog = MaterialAlertDialogBuilder(requireContext())
+                    .setTitle(R.string.dialog_joblist_text_title)
+                    .setMessage(R.string.dialog_joblist_text_message)
+                    .setPositiveButton(R.string.dialog_joblist_positive_title) { dialog, which ->
+                        FragmentNavigationUtils.addFragmentOpenWithOneStringBundleValueAndSourceFragment(
+                            fragmentManager = parentFragmentManager,
+                            fragment = JobApplyToFragmentView(),
+                            argumentKey = "jobId",
+                            argumentValue = string,
+                            argumentSourceFragmentName = "ProfileJobListFragmentView"
+                        )
+                    }
+                    .setNeutralButton(R.string.dialog_joblist_neutral_title) { dialog, which ->
+                        dialog.dismiss()
+                    }
+                    .setNegativeButton(R.string.dialog_joblist_negative_title) { dialog, which ->
+                        viewModelProfileJobList.deleteJob(string)
+                    }
 
-            if(!SessionManager.getIsAdmin(requireContext())) {
-                materialDialog.setNegativeButton("") { dialog, which -> }
+                if(!SessionManager.getIsAdmin(requireContext())) {
+                    materialDialog.setNegativeButton("") { dialog, which -> }
+                }
+                materialDialog.show()
             }
-            materialDialog.show()
         }
     }
